test: add perfomance benchmarking to test suite
