+++ package.json
@@ -17,12 +17,13 @@
   },
   "license": "BSD-2-Clause",
   "devDependencies": {
+    "@types/node": "12.12.21",
+    "benchmark": "^2.1.4",
     "mocha": "6.2.2",
     "phantomjs": "2.1.7",
     "sinon": "9.0.1",
-    "uglify-js": "3.8.0",
     "typescript": "3.8.3",
-    "@types/node": "12.12.21"
+    "uglify-js": "3.8.0"
   },
   "peerDependencies": {
     "typescript": "^3.6.3",
